Move RecommendationMetadataType and RecommendationDataType interfaces to a new file
